refactor & feat: Cleaned up messy code and added repository to store the data

+++ internal/migrations/20251013105513_create_inventory_schema.sql
@@ -4,8 +4,8 @@ CREATE TABLE inventory (
     item TEXT,
     quantity INTEGER,
     unit TEXT,
-    wholesale_price_per_quantity INTEGER,
-    total_cost_of_product INTEGER,
+    wholesale_price_per_quantity DECIMAL,
+    total_cost_of_product DECIMAL,
     CHECK (unit IN ('kg', 'g', 'dozen', 'unit'))
 );
 -- +goose StatementEnd
